+++ .editorconfig
@@ -1,0 +1,15 @@
+root = true
+
+[*]
+insert_final_newline = true
+
+[*.{js}]
+charset = utf-8
+
+[src/**.js]
+indent_style = space
+indent_size = 2
+
+[{package.json,.travis.yml}]
+indent_style = space
+indent_size = 2
